Add folder for docs to be OCRed w/ folder watcher
Using chokidar, the OCRManager monitors a folder. For each
file added to the folder, it runs the file through
pdfsearchify.

On startup, ORCManager now requires a parameter with a path to
a valid JSON file. This file must contain a key for 'watchpath'
and 'finishpath', each with a value of a valid path. The
'watchpath' is the directory to monitor for PDFs to OCR, and
the 'finishpath' is where the OCRed docs are deposited.