Add file move on OCR error/completion
OCRManager now creates the necessary subdirectory
tree in the finishpath for every file, so that
files maintain the same relative directory path in
the finishpath and errorpath that they have in the
watchpath.

On notification of successful ORC, OCRManager checks
to make sure that the OCRed file does exist in the
finishpath and, if found, removed the file from the
watchpath.

On notification of an error in the OCR process,
OCRManager moves the file to the currect subdirectory
in the errorpath. If the move is successful, OCRManager
removes the file from the watchpath.