Add processpath and use usePolling for chokidar
A new key for "processfile" has been added as a required
key to the settings JSON file. The value of this key must
be a valid folder path, which will be used for staging
the PDF documents during the OCR process. Before processing,
PDFs are moved to this folder and, on success, removed from
the watchpath.

The Chokidar folder watcher now sets the usePolling option
to true,