Refactor to add hocr to original pdf
Before, the hocr-pdf script drew the .pnm file created by
unpaper onto a ReportLab canvas, and drew the PDF text
read from the hocr file on top of the canvas. Now, hocr-pdf
draws the hocr text onto a blank canvas and then merges that
ReportLab pdf (using pyPdf) with the original pdf.

Using a 345kB pdf, the old method resulted in an OCRed pdf
size of 815kB. The new methods results in an OCRed pdf size
of 375kB.