Refactor chain to separate original pdf
Before, each page of the PDF was extracted as a pdf, and the
deskew function outputted a pdf file. A copy of this file was
then rasterized for OCR, and the pdf was later used to merge
the OCRed HOCR file. Now, the page is extracted as both a pdf
and pnm file. Both files are deskewed, the pnm file is send
through the ocr process, and the pdf is used to merge the
HOCR file. This improved the accuracy of the OCR