Refactor OCRManager to use async.js queues
OCRManager now uses async.js to enforce orderly
callback structure and to iterate through the
queue of files created by the chokidar folder
watcher.

An option for 'keeppath' has been added, which will
move a copy of every file to the specified path before
attempting OCR.

An option for 'workers' has been added, which accepts
and int and sets the number of workers to be utilized in
the async.js queue to OCR files.

An option for 'pdfsearchify' has been added, which
accepts a JSON object to pass to pdfsearchify as its
options.