Refactor OCRManager
OCRManager has been refactored as a node module in
./ocrmanager/index.js. The original OCRManager.js has been
modified to accept a settings filepath and start the ocrmanager
module with this filepath.

The Winston logging module has been added to OCRManager. The
settings JSON file now accepts two additional, optional arguments.
The 'filelog' parameter accepts a boolean argument which, when
true, logs all Winston logs to a file in ./logs/. The 'loglevel'
argument accepts a string that, when set to 'verbose', logs all
'error', 'warn', and 'info' Winston logs in both the logfile and
in the console. If the loglevel is not set, Winston only logs
'error' and 'warn' statements.